[proposals] Move Helper to Vote, remove proposals from summary
