+++ src/Badger/GameBundle/Controller/BadgeProposalController.php
@@ -28,10 +28,12 @@ class BadgeProposalController extends DefaultController
      */
     public function indexAction(Request $request)
     {
+        $badgeProposalRepository = $this->get('badger.game.repository.badge_proposal');
         $badgeVoteSummaryFactory = $this->get('badger.game.badge_vote_summary.factory');
         $badgeVoteSummary = $badgeVoteSummaryFactory->create($this->getUser());
 
         return $this->render('@Game/badge-proposals/index.html.twig', [
+            'badgeProposals'   => $badgeProposalRepository->findAllSorted(),
             'badgeVoteSummary' => $badgeVoteSummary,
         ]);
     }
