+++ src/Badger/GameBundle/DependencyInjection/GameExtension.php
@@ -21,7 +21,7 @@ class GameExtension extends Extension
     {
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('factories.yml');
-        $loader->load('helpers.yml');
+        $loader->load('votes.yml');
         $loader->load('removers.yml');
         $loader->load('repositories.yml');
         $loader->load('twig.yml');
