+++ src/Badger/GameBundle/Resources/views/badge-proposals/index.html.twig
@@ -9,7 +9,6 @@
             </h2>
 
             <div class="badge-proposals">
-                {% set badgeProposals = badgeVoteSummary.getBadgeProposals %}
                 {% if badgeProposals is not empty %}
                     {% for badgeProposal in badgeProposals %}
                         <div
